removed SubjectNode type
Kept to using any as it's contained in the subject graph. react force
graph has no way of modifying the return type of the node object. Can
possibly protect against this using type guards, but keeping track of
the types strictly in this component isn't too difficult so will leave
it for now.