feat: finished search backend and working on front
Back-end route searches for words in subject name, resource name, and resource
notes that starts with the given input.

Refactored selections in search-input to be ReactNodes to have more
options in how selections are displayed. Created a filter-search-input
to take selection strings and handle filter logic for searching
selections that start with the typed in value.

Finished the vault-search component front-end and now connecting it with
the back-end search.